requirements.txt
asgiref==3.4.1
boto3==1.19.0
botocore==1.22.0
dj-database-url==0.5.0
Django==3.2.7
django-debug-toolbar==3.2.2
django-ses==2.3.0
django-storages==1.12.2
djangorestframework==3.12.4
future==0.18.2
jmespath==0.10.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
Pillow==9.0.0